backend/requirements.txt
# -----------------------------
# Core API Framework
# -----------------------------
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.6.4
pydantic-settings==2.2.1
requests==2.31.0
python-dotenv==1.0.1

# -----------------------------
# NLP & Embeddings
# -----------------------------
spacy==3.7.2
sentence-transformers==2.5.1
rapidfuzz==3.6.1
scikit-learn==1.4.1.post1
numpy==1.26.4

# -----------------------------
# PDF, DOC Parsing
# -----------------------------
pdfplumber==0.11.0
pdfminer.six==20231228
PyPDF2==3.0.1
python-docx==1.1.0
docx2txt==0.8
python-magic==0.4.27

# -----------------------------
# Deep Learning (CPU)
# -----------------------------
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0

# -----------------------------
# Vector Database (FAISS)
# -----------------------------
faiss-cpu==1.7.4

# -----------------------------
# Data
# -----------------------------
pandas==2.2.1

# -----------------------------
# LangChain + Gemini (LLM Learning Path)
# -----------------------------
langchain==0.2.3
langchain-core==0.2.3
langchain-community==0.2.3
google-generativeai==0.5.2
langchain-google-genai==1.0.5

# -----------------------------
# Utility
# -----------------------------
tqdm==4.66.2
python-dotenv
pymongo